Read an enum's discriminants as rustc computes them (#265)
`VariantDiscr::Relative(i)` is the distance from the nearest preceding explicit
discriminant, not a discriminant, so every variant following an explicit one was
encoded with a value that the `SwitchInt` of a `match` never compares against:
its arm ran under an unsatisfiable path assumption and verified vacuously,
accepting programs that panic. Reading an explicit discriminant went through
`ScalarInt::to_u32`, which asserts a four-byte tag, so an enum carrying one bugs
out on every tag type but `i32`/`u32` — `isize` by default, and `i8` for
`core::cmp::Ordering`, which any `#[derive(PartialOrd)]` in the crate pulls in.

Take the absolute discriminants rustc already computes, sign-extended to the
width their tag type declares so they agree with the `SwitchInt` targets, and
carry them as `i64` from there on.

Fixes #126

+++ src/chc.rs
@@ -1948,7 +1948,7 @@ pub struct DatatypeSelector {
 pub struct DatatypeCtor {
     pub symbol: DatatypeSymbol,
     pub selectors: Vec<DatatypeSelector>,
-    pub discriminant: u32,
+    pub discriminant: i64,
 }
 
 /// A datatype definition.
